[runtime-security] delete old test file
