+++ pkg/security/tests/chown32_test.go
@@ -3,28 +3,30 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-// +build functionaltests,386
+// +build functionaltests,!386
 
 package tests
 
 import (
 	"os"
-	"syscall"
+	"os/exec"
+	"path/filepath"
 	"testing"
-	"time"
 
 	"github.com/DataDog/datadog-agent/pkg/security/rules"
+	"github.com/DataDog/datadog-agent/pkg/security/tests/syscall_tester"
+	"gotest.tools/assert"
 )
 
-func TestChown(t *testing.T) {
+func TestChown32(t *testing.T) {
 	ruleDef := &rules.RuleDefinition{
 		ID:         "test_rule",
-		Expression: `chown.file.path == "{{.Root}}/test-chown" && chown.file.destination.uid in [100, 101, 102, 103] && chown.file.destination.gid in [200, 201, 202, 203]`,
+		Expression: `chown.file.path == "{{.Root}}/test-chown" && chown.file.destination.uid in [100, 101, 102, 103, 104, 105, 106] && chown.file.destination.gid in [200, 201, 202, 203, 204, 205, 206]`,
 	}
 
 	ruleDef2 := &rules.RuleDefinition{
 		ID:         "test_rule2",
-		Expression: `chown.file.path == "{{.Root}}/test-symlink" && chown.file.destination.uid in [100, 101, 102, 103] && chown.file.destination.gid in [200, 201, 202, 203]`,
+		Expression: `chown.file.path == "{{.Root}}/test-symlink" && chown.file.destination.uid in [100, 101, 102, 103, 104, 105, 106] && chown.file.destination.gid in [200, 201, 202, 203, 204, 205, 206]`,
 	}
 
 	test, err := newTestModule(nil, []*rules.RuleDefinition{ruleDef, ruleDef2}, testOpts{})
@@ -33,57 +35,49 @@ func TestChown(t *testing.T) {
 	}
 	defer test.Close()
 
+	syscallTester, err := loadSyscallTester(t)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	prevUID := 98
+	prevGID := 99
 	fileMode := 0o447
-	expectedMode := applyUmask(fileMode)
-	testFile, testFilePtr, err := test.CreateWithOptions("test-chown", 98, 99, fileMode)
+	expectedMode := uint32(applyUmask(fileMode))
+	testFile, _, err := test.CreateWithOptions("test-chown", 98, 99, fileMode)
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	t.Run("chown", func(t *testing.T) {
-		if _, _, errno := syscall.Syscall(syscall.SYS_CHOWN, uintptr(testFilePtr), 100, 200); errno != 0 {
-			t.Fatal(err)
+		sideTester := exec.Command(syscallTester, "chown", testFile, "100", "200")
+		if output, err := sideTester.CombinedOutput(); err != nil {
+			t.Error(string(output))
+			t.Error(err)
 		}
+
+		defer func() {
+			prevUID = 100
+			prevGID = 200
+		}()
 
 		event, _, err := test.GetEvent()
 		if err != nil {
 			t.Error(err)
 		} else {
-			if event.GetType() != "chown" {
-				t.Errorf("expected chown event, got %s", event.GetType())
-			}
+			assert.Equal(t, event.GetType(), "chown", "wrong event type")
+			assert.Equal(t, event.Chown.UID, uint32(100), "wrong user")
+			assert.Equal(t, event.Chown.GID, uint32(200), "wrong user")
+			assert.Equal(t, event.Chown.File.Inode, getInode(t, testFile), "wrong inode")
+			assertRights(t, event.Chown.File.Mode, uint16(expectedMode), "wrong initial mode")
+			assert.Equal(t, event.Chown.File.UID, uint32(prevUID), "wrong initial user")
+			assert.Equal(t, event.Chown.File.GID, uint32(prevGID), "wrong initial group")
 
-			if user := event.Chown.UID; user != 100 {
-				t.Errorf("expected chown user 100, got %d", user)
-			}
+			assertNearTime(t, event.Chown.File.MTime)
+			assertNearTime(t, event.Chown.File.CTime)
 
-			if group := event.Chown.GID; group != 200 {
-				t.Errorf("expected chown group 200, got %d", group)
-			}
-
-			if inode := getInode(t, testFile); inode != event.Chown.File.Inode {
-				t.Logf("expected inode %d, got %d", event.Chown.File.Inode, inode)
-			}
-
-			if int(event.Chown.File.Mode)&expectedMode != expectedMode {
-				t.Errorf("expected initial mode %d, got %d", expectedMode, int(event.Chown.File.Mode)&expectedMode)
-			}
-
-			if event.Chown.File.UID != 98 {
-				t.Errorf("expected initial UID %d, got %d", 98, event.Chown.File.UID)
-			}
-
-			if event.Chown.File.GID != 99 {
-				t.Errorf("expected initial GID %d, got %d", 99, event.Chown.File.GID)
-			}
-
-			now := time.Now()
-			if event.Chown.File.MTime.After(now) || event.Chown.File.MTime.Before(now.Add(-1*time.Hour)) {
-				t.Errorf("expected mtime close to %s, got %s", now, event.Chown.File.MTime)
-			}
-
-			if event.Chown.File.CTime.After(now) || event.Chown.File.CTime.Before(now.Add(-1*time.Hour)) {
-				t.Errorf("expected ctime close to %s, got %s", now, event.Chown.File.CTime)
+			if testEnvironment == DockerEnvironment {
+				testContainerPath(t, event, "chown.file.container_path")
 			}
 
 			if !validateChownSchema(t, event) {
@@ -93,54 +87,34 @@ func TestChown(t *testing.T) {
 	})
 
 	t.Run("fchown", func(t *testing.T) {
-		f, err := os.Open(testFile)
-		if err != nil {
-			t.Fatal(err)
+		sideTester := exec.Command(syscallTester, "fchown", testFile, "101", "201")
+		if output, err := sideTester.CombinedOutput(); err != nil {
+			t.Error(string(output))
+			t.Error(err)
 		}
-		// fchown syscall
-		if _, _, errno := syscall.Syscall(syscall.SYS_FCHOWN, f.Fd(), 101, 201); errno != 0 {
-			t.Fatal(err)
-		}
-		defer f.Close()
+
+		defer func() {
+			prevUID = 101
+			prevGID = 201
+		}()
 
 		event, _, err := test.GetEvent()
 		if err != nil {
 			t.Error(err)
 		} else {
-			if event.GetType() != "chown" {
-				t.Errorf("expected chown event, got %s", event.GetType())
-			}
+			assert.Equal(t, event.GetType(), "chown", "wrong event type")
+			assert.Equal(t, event.Chown.UID, uint32(101), "wrong user")
+			assert.Equal(t, event.Chown.GID, uint32(201), "wrong user")
+			assert.Equal(t, event.Chown.File.Inode, getInode(t, testFile), "wrong inode")
+			assertRights(t, event.Chown.File.Mode, uint16(expectedMode), "wrong initial mode")
+			assert.Equal(t, event.Chown.File.UID, uint32(prevUID), "wrong initial user")
+			assert.Equal(t, event.Chown.File.GID, uint32(prevGID), "wrong initial group")
 
-			if user := event.Chown.UID; user != 101 {
-				t.Errorf("expected chown user 101, got %d", user)
-			}
+			assertNearTime(t, event.Chown.File.MTime)
+			assertNearTime(t, event.Chown.File.CTime)
 
-			if group := event.Chown.GID; group != 201 {
-				t.Errorf("expected chown group 201, got %d", group)
-			}
-			if inode := getInode(t, testFile); inode != event.Chown.File.Inode {
-				t.Logf("expected inode %d, got %d", event.Chown.File.Inode, inode)
-			}
-
-			if int(event.Chown.File.Mode)&expectedMode != expectedMode {
-				t.Errorf("expected initial mode %d, got %d", expectedMode, int(event.Chown.File.Mode)&expectedMode)
-			}
-
-			if event.Chown.File.UID != 100 {
-				t.Errorf("expected initial UID %d, got %d", 100, event.Chown.File.UID)
-			}
-
-			if event.Chown.File.GID != 200 {
-				t.Errorf("expected initial GID %d, got %d", 200, event.Chown.File.GID)
-			}
-
-			now := time.Now()
-			if event.Chown.File.MTime.After(now) || event.Chown.File.MTime.Before(now.Add(-1*time.Hour)) {
-				t.Errorf("expected mtime close to %s, got %s", now, event.Chown.File.MTime)
-			}
-
-			if event.Chown.File.CTime.After(now) || event.Chown.File.CTime.Before(now.Add(-1*time.Hour)) {
-				t.Errorf("expected ctime close to %s, got %s", now, event.Chown.File.CTime)
+			if testEnvironment == DockerEnvironment {
+				testContainerPath(t, event, "chown.file.container_path")
 			}
 
 			if !validateChownSchema(t, event) {
@@ -150,49 +124,34 @@ func TestChown(t *testing.T) {
 	})
 
 	t.Run("fchownat", func(t *testing.T) {
-		if _, _, errno := syscall.Syscall6(syscall.SYS_FCHOWNAT, 0, uintptr(testFilePtr), uintptr(102), uintptr(202), 0x100, 0); errno != 0 {
-			t.Fatal(err)
+		sideTester := exec.Command(syscallTester, "fchownat", testFile, "102", "202")
+		if output, err := sideTester.CombinedOutput(); err != nil {
+			t.Error(string(output))
+			t.Error(err)
 		}
+
+		defer func() {
+			prevUID = 102
+			prevGID = 202
+		}()
 
 		event, _, err := test.GetEvent()
 		if err != nil {
 			t.Error(err)
 		} else {
-			if event.GetType() != "chown" {
-				t.Errorf("expected chown event, got %s", event.GetType())
-			}
+			assert.Equal(t, event.GetType(), "chown", "wrong event type")
+			assert.Equal(t, event.Chown.UID, uint32(102), "wrong user")
+			assert.Equal(t, event.Chown.GID, uint32(202), "wrong user")
+			assert.Equal(t, event.Chown.File.Inode, getInode(t, testFile), "wrong inode")
+			assertRights(t, event.Chown.File.Mode, uint16(expectedMode), "wrong initial mode")
+			assert.Equal(t, event.Chown.File.UID, uint32(prevUID), "wrong initial user")
+			assert.Equal(t, event.Chown.File.GID, uint32(prevGID), "wrong initial group")
 
-			if user := event.Chown.UID; user != 102 {
-				t.Errorf("expected chown user 102, got %d", user)
-			}
+			assertNearTime(t, event.Chown.File.MTime)
+			assertNearTime(t, event.Chown.File.CTime)
 
-			if group := event.Chown.GID; group != 202 {
-				t.Errorf("expected chown group 202, got %d", group)
-			}
-
-			if inode := getInode(t, testFile); inode != event.Chown.File.Inode {
-				t.Logf("expected inode %d, got %d", event.Chown.File.Inode, inode)
-			}
-
-			if int(event.Chown.File.Mode)&expectedMode != expectedMode {
-				t.Errorf("expected initial mode %d, got %d", expectedMode, int(event.Chown.File.Mode)&expectedMode)
-			}
-
-			if event.Chown.File.UID != 101 {
-				t.Errorf("expected initial UID %d, got %d", 101, event.Chown.File.UID)
-			}
-
-			if event.Chown.File.GID != 201 {
-				t.Errorf("expected initial GID %d, got %d", 201, event.Chown.File.GID)
-			}
-
-			now := time.Now()
-			if event.Chown.File.MTime.After(now) || event.Chown.File.MTime.Before(now.Add(-1*time.Hour)) {
-				t.Errorf("expected mtime close to %s, got %s", now, event.Chown.File.MTime)
-			}
-
-			if event.Chown.File.CTime.After(now) || event.Chown.File.CTime.Before(now.Add(-1*time.Hour)) {
-				t.Errorf("expected ctime close to %s, got %s", now, event.Chown.File.CTime)
+			if testEnvironment == DockerEnvironment {
+				testContainerPath(t, event, "chown.file.container_path")
 			}
 
 			if !validateChownSchema(t, event) {
@@ -202,7 +161,7 @@ func TestChown(t *testing.T) {
 	})
 
 	t.Run("lchown", func(t *testing.T) {
-		testSymlink, testSymlinkPtr, err := test.Path("test-symlink")
+		testSymlink, _, err := test.Path("test-symlink")
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -212,53 +171,29 @@ func TestChown(t *testing.T) {
 		}
 		defer os.Remove(testSymlink)
 
-		if _, _, errno := syscall.Syscall(syscall.SYS_LCHOWN, uintptr(testSymlinkPtr), uintptr(103), uintptr(203)); errno != 0 {
-			t.Fatal(err)
+		sideTester := exec.Command(syscallTester, "lchown", testSymlink, "103", "203")
+		if output, err := sideTester.CombinedOutput(); err != nil {
+			t.Error(string(output))
+			t.Error(err)
 		}
 
-		event, rule, err := test.GetEvent()
+		event, _, err := test.GetEvent()
 		if err != nil {
 			t.Error(err)
 		} else {
-			if event.GetType() != "chown" {
-				t.Errorf("expected chown event, got %s", event.GetType())
-			}
+			assert.Equal(t, event.GetType(), "chown", "wrong event type")
+			assert.Equal(t, event.Chown.UID, uint32(103), "wrong user")
+			assert.Equal(t, event.Chown.GID, uint32(203), "wrong user")
+			assert.Equal(t, event.Chown.File.Inode, getInode(t, testSymlink), "wrong inode")
+			assertRights(t, event.Chown.File.Mode, uint16(0o777), "wrong initial mode")
+			assert.Equal(t, event.Chown.File.UID, uint32(0), "wrong initial user")
+			assert.Equal(t, event.Chown.File.GID, uint32(0), "wrong initial group")
 
-			if rule.ID != "test_rule2" {
-				t.Errorf("expected triggered rule test_rule2, got %s", rule.ID)
-			}
+			assertNearTime(t, event.Chown.File.MTime)
+			assertNearTime(t, event.Chown.File.CTime)
 
-			if user := event.Chown.UID; user != 103 {
-				t.Errorf("expected chown user 103, got %d", user)
-			}
-
-			if group := event.Chown.GID; group != 203 {
-				t.Errorf("expected chown group 203, got %d", group)
-			}
-
-			if inode := getInode(t, testSymlink); inode != event.Chown.File.Inode {
-				t.Logf("expected inode %d, got %d", event.Chown.File.Inode, inode)
-			}
-
-			if int(event.Chown.File.Mode)&0o777 != 0o777 {
-				t.Errorf("expected initial mode %d, got %d", 0o777, int(event.Chown.File.Mode))
-			}
-
-			if event.Chown.File.UID != 0 {
-				t.Errorf("expected initial UID %d, got %d", 0, event.Chown.File.UID)
-			}
-
-			if event.Chown.File.GID != 0 {
-				t.Errorf("expected initial GID %d, got %d", 0, event.Chown.File.GID)
-			}
-
-			now := time.Now()
-			if event.Chown.File.MTime.After(now) || event.Chown.File.MTime.Before(now.Add(-1*time.Hour)) {
-				t.Errorf("expected mtime close to %s, got %s", now, event.Chown.File.MTime)
-			}
-
-			if event.Chown.File.CTime.After(now) || event.Chown.File.CTime.Before(now.Add(-1*time.Hour)) {
-				t.Errorf("expected ctime close to %s, got %s", now, event.Chown.File.CTime)
+			if testEnvironment == DockerEnvironment {
+				testContainerPath(t, event, "chown.file.container_path")
 			}
 
 			if !validateChownSchema(t, event) {
@@ -268,7 +203,7 @@ func TestChown(t *testing.T) {
 	})
 
 	t.Run("lchown32", func(t *testing.T) {
-		testSymlink, testSymlinkPtr, err := test.Path("test-symlink")
+		testSymlink, _, err := test.Path("test-symlink")
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -278,53 +213,29 @@ func TestChown(t *testing.T) {
 		}
 		defer os.Remove(testSymlink)
 
-		if _, _, errno := syscall.Syscall(syscall.SYS_LCHOWN32, uintptr(testSymlinkPtr), uintptr(103), uintptr(203)); errno != 0 {
-			t.Fatal(err)
+		sideTester := exec.Command(syscallTester, "lchown32", testSymlink, "104", "204")
+		if output, err := sideTester.CombinedOutput(); err != nil {
+			t.Error(string(output))
+			t.Error(err)
 		}
 
-		event, rule, err := test.GetEvent()
+		event, _, err := test.GetEvent()
 		if err != nil {
 			t.Error(err)
 		} else {
-			if event.GetType() != "chown" {
-				t.Errorf("expected chown event, got %s", event.GetType())
-			}
+			assert.Equal(t, event.GetType(), "chown", "wrong event type")
+			assert.Equal(t, event.Chown.UID, uint32(104), "wrong user")
+			assert.Equal(t, event.Chown.GID, uint32(204), "wrong user")
+			assert.Equal(t, event.Chown.File.Inode, getInode(t, testSymlink), "wrong inode")
+			assertRights(t, event.Chown.File.Mode, uint16(0o777), "wrong initial mode")
+			assert.Equal(t, event.Chown.File.UID, uint32(0), "wrong initial user")
+			assert.Equal(t, event.Chown.File.GID, uint32(0), "wrong initial group")
 
-			if rule.ID != "test_rule2" {
-				t.Errorf("expected triggered rule test_rule2, got %s", rule.ID)
-			}
+			assertNearTime(t, event.Chown.File.MTime)
+			assertNearTime(t, event.Chown.File.CTime)
 
-			if user := event.Chown.UID; user != 103 {
-				t.Errorf("expected chown user 103, got %d", user)
-			}
-
-			if group := event.Chown.GID; group != 203 {
-				t.Errorf("expected chown group 203, got %d", group)
-			}
-
-			if inode := getInode(t, testSymlink); inode != event.Chown.File.Inode {
-				t.Logf("expected inode %d, got %d", event.Chown.File.Inode, inode)
-			}
-
-			if int(event.Chown.File.Mode)&0o777 != 0o777 {
-				t.Errorf("expected initial mode %d, got %d", 0o777, int(event.Chown.File.Mode))
-			}
-
-			if event.Chown.File.UID != 0 {
-				t.Errorf("expected initial UID %d, got %d", 0, event.Chown.File.UID)
-			}
-
-			if event.Chown.File.GID != 0 {
-				t.Errorf("expected initial GID %d, got %d", 0, event.Chown.File.GID)
-			}
-
-			now := time.Now()
-			if event.Chown.File.MTime.After(now) || event.Chown.File.MTime.Before(now.Add(-1*time.Hour)) {
-				t.Errorf("expected mtime close to %s, got %s", now, event.Chown.File.MTime)
-			}
-
-			if event.Chown.File.CTime.After(now) || event.Chown.File.CTime.Before(now.Add(-1*time.Hour)) {
-				t.Errorf("expected ctime close to %s, got %s", now, event.Chown.File.CTime)
+			if testEnvironment == DockerEnvironment {
+				testContainerPath(t, event, "chown.file.container_path")
 			}
 
 			if !validateChownSchema(t, event) {
@@ -334,55 +245,34 @@ func TestChown(t *testing.T) {
 	})
 
 	t.Run("fchown32", func(t *testing.T) {
-		f, err := os.Open(testFile)
-		if err != nil {
-			t.Fatal(err)
+		sideTester := exec.Command(syscallTester, "fchown32", testFile, "105", "205")
+		if output, err := sideTester.CombinedOutput(); err != nil {
+			t.Error(string(output))
+			t.Error(err)
 		}
-		// fchown syscall
-		if _, _, errno := syscall.Syscall(syscall.SYS_FCHOWN32, f.Fd(), 101, 201); errno != 0 {
-			t.Fatal(err)
-		}
-		defer f.Close()
+
+		defer func() {
+			prevUID = 105
+			prevGID = 205
+		}()
 
 		event, _, err := test.GetEvent()
 		if err != nil {
 			t.Error(err)
 		} else {
-			if event.GetType() != "chown" {
-				t.Errorf("expected chown event, got %s", event.GetType())
-			}
+			assert.Equal(t, event.GetType(), "chown", "wrong event type")
+			assert.Equal(t, event.Chown.UID, uint32(105), "wrong user")
+			assert.Equal(t, event.Chown.GID, uint32(205), "wrong user")
+			assert.Equal(t, event.Chown.File.Inode, getInode(t, testFile), "wrong inode")
+			assertRights(t, event.Chown.File.Mode, uint16(expectedMode), "wrong initial mode")
+			assert.Equal(t, event.Chown.File.UID, uint32(prevUID), "wrong initial user")
+			assert.Equal(t, event.Chown.File.GID, uint32(prevGID), "wrong initial group")
 
-			if user := event.Chown.UID; user != 101 {
-				t.Errorf("expected chown user 101, got %d", user)
-			}
+			assertNearTime(t, event.Chown.File.MTime)
+			assertNearTime(t, event.Chown.File.CTime)
 
-			if group := event.Chown.GID; group != 201 {
-				t.Errorf("expected chown group 201, got %d", group)
-			}
-
-			if inode := getInode(t, testFile); inode != event.Chown.File.Inode {
-				t.Logf("expected inode %d, got %d", event.Chown.File.Inode, inode)
-			}
-
-			if int(event.Chown.File.Mode)&expectedMode != expectedMode {
-				t.Errorf("expected initial mode %d, got %d", expectedMode, int(event.Chown.File.Mode)&expectedMode)
-			}
-
-			if event.Chown.File.UID != 102 {
-				t.Errorf("expected initial UID %d, got %d", 102, event.Chown.File.UID)
-			}
-
-			if event.Chown.File.GID != 202 {
-				t.Errorf("expected initial GID %d, got %d", 202, event.Chown.File.GID)
-			}
-
-			now := time.Now()
-			if event.Chown.File.MTime.After(now) || event.Chown.File.MTime.Before(now.Add(-1*time.Hour)) {
-				t.Errorf("expected mtime close to %s, got %s", now, event.Chown.File.MTime)
-			}
-
-			if event.Chown.File.CTime.After(now) || event.Chown.File.CTime.Before(now.Add(-1*time.Hour)) {
-				t.Errorf("expected ctime close to %s, got %s", now, event.Chown.File.CTime)
+			if testEnvironment == DockerEnvironment {
+				testContainerPath(t, event, "chown.file.container_path")
 			}
 
 			if !validateChownSchema(t, event) {
@@ -392,49 +282,34 @@ func TestChown(t *testing.T) {
 	})
 
 	t.Run("chown32", func(t *testing.T) {
-		if _, _, errno := syscall.Syscall(syscall.SYS_CHOWN32, uintptr(testFilePtr), 100, 200); errno != 0 {
-			t.Fatal(err)
+		sideTester := exec.Command(syscallTester, "chown32", testFile, "106", "206")
+		if output, err := sideTester.CombinedOutput(); err != nil {
+			t.Error(string(output))
+			t.Error(err)
 		}
+
+		defer func() {
+			prevUID = 106
+			prevGID = 206
+		}()
 
 		event, _, err := test.GetEvent()
 		if err != nil {
 			t.Error(err)
 		} else {
-			if event.GetType() != "chown" {
-				t.Errorf("expected chown event, got %s", event.GetType())
-			}
+			assert.Equal(t, event.GetType(), "chown", "wrong event type")
+			assert.Equal(t, event.Chown.UID, uint32(106), "wrong user")
+			assert.Equal(t, event.Chown.GID, uint32(206), "wrong user")
+			assert.Equal(t, event.Chown.File.Inode, getInode(t, testFile), "wrong inode")
+			assertRights(t, event.Chown.File.Mode, uint16(expectedMode), "wrong initial mode")
+			assert.Equal(t, event.Chown.File.UID, uint32(prevUID), "wrong initial user")
+			assert.Equal(t, event.Chown.File.GID, uint32(prevGID), "wrong initial group")
 
-			if user := event.Chown.UID; user != 100 {
-				t.Errorf("expected chown user 100, got %d", user)
-			}
+			assertNearTime(t, event.Chown.File.MTime)
+			assertNearTime(t, event.Chown.File.CTime)
 
-			if group := event.Chown.GID; group != 200 {
-				t.Errorf("expected chown group 200, got %d", group)
-			}
-
-			if inode := getInode(t, testFile); inode != event.Chown.File.Inode {
-				t.Logf("expected inode %d, got %d", event.Chown.File.Inode, inode)
-			}
-
-			if int(event.Chown.File.Mode)&expectedMode != expectedMode {
-				t.Errorf("expected initial mode %d, got %d", expectedMode, int(event.Chown.File.Mode)&expectedMode)
-			}
-
-			if event.Chown.File.UID != 101 {
-				t.Errorf("expected initial UID %d, got %d", 101, event.Chown.File.UID)
-			}
-
-			if event.Chown.File.GID != 201 {
-				t.Errorf("expected initial GID %d, got %d", 201, event.Chown.File.GID)
-			}
-
-			now := time.Now()
-			if event.Chown.File.MTime.After(now) || event.Chown.File.MTime.Before(now.Add(-1*time.Hour)) {
-				t.Errorf("expected mtime close to %s, got %s", now, event.Chown.File.MTime)
-			}
-
-			if event.Chown.File.CTime.After(now) || event.Chown.File.CTime.Before(now.Add(-1*time.Hour)) {
-				t.Errorf("expected ctime close to %s, got %s", now, event.Chown.File.CTime)
+			if testEnvironment == DockerEnvironment {
+				testContainerPath(t, event, "chown.file.container_path")
 			}
 
 			if !validateChownSchema(t, event) {
@@ -442,4 +317,31 @@ func TestChown(t *testing.T) {
 			}
 		}
 	})
+}
+
+func loadSyscallTester(t *testing.T) (string, error) {
+	testerBin, err := syscall_tester.Asset("/syscall_x86_tester")
+	if err != nil {
+		return "", err
+	}
+
+	binPath := filepath.Join(t.TempDir(), "syscall_x86_tester")
+	f, err := os.Create(binPath)
+	if err != nil {
+		return "", err
+	}
+
+	if _, err = f.Write(testerBin); err != nil {
+		return "", err
+	}
+
+	if err := f.Close(); err != nil {
+		return "", err
+	}
+
+	if err := os.Chmod(binPath, 0o700); err != nil {
+		return "", err
+	}
+
+	return binPath, nil
 }
